Get player to connect to host
