+++ src/components/App.tsx
@@ -1,13 +1,10 @@
-import React, { useState } from "react";
+import React from "react";
 import styled from "styled-components";
-import { Box, Button, Divider, Stack } from "@chakra-ui/react";
-import { Link as RouterLink, Route } from "wouter";
-import PeerConnectionManager from "./peer/PeerConnectionManager";
+import { Box } from "@chakra-ui/react";
+import { Route } from "wouter";
 import MainMenu from "./pages/MainMenu";
 import Host from "./pages/Host";
 import Join from "./pages/Join";
-
-const namespace = "sfqxWNtbOKrs45NDDZRvOaP4vAApMTc1";
 
 const AppContainer = styled.div`
   background-color: #282c34;
@@ -20,19 +17,6 @@ const AppContainer = styled.div`
 `;
 
 function App() {
-  const [connected, setConnected] = useState(false);
-  const [isConnecting, setIsConnecting] = useState(false);
-
-  const handleConnect = (roomCode: string, name: string) => {
-    setIsConnecting(true);
-    PeerConnectionManager.register(`${namespace} ${roomCode} ${name}`)
-      .then(() => {
-        setConnected(true);
-      })
-      .catch((err) => console.error(err))
-      .finally(() => setIsConnecting(false));
-  };
-
   return (
     <AppContainer>
       <Box bg="white" color="black" p={4} mb={60} w={275} boxShadow="xl">
