+++ src/config.js
@@ -1,0 +1,1 @@
+export const namespace = "sfqxWNtbOKrs45NDDZRvOaP4vAApMTc1";
